ServerlessComputing/GoogleCloud/src/resize/requirements.txt
pillow==6.2.1
flask==1.1.1